add test for stream_and_receive_listener
